# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Portfolio manager API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/kaseat/pManager/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "totallink.ru",
    "basePath": "/api",
    "paths": {
        "/misc/gmail/url": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets url for GMail auth",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get GMail auth url",
                "operationId": "get-gmail-url",
                "responses": {
                    "200": {
                        "description": "Returns url for GMail auth",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.gmailAuthURLSuccess"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/misc/sync/price": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sync prices for given portfolio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc"
                ],
                "summary": "Sync prices",
                "operationId": "sync-price",
                "responses": {
                    "200": {
                        "description": "Returns success status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.commonResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/portfolios/{id}/average": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets average price of given ticker",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc"
                ],
                "summary": "Get average",
                "operationId": "get-average",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ticker",
                        "name": "ticker",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns average price of given ticker",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.getAverageSuccess"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/portfolios/{id}/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets balance of given currency",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc"
                ],
                "summary": "Get balance",
                "operationId": "get-balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Currency",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "On date",
                        "name": "on",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns balance of given currency",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.getBalanceSuccess"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/portfolios/{id}/sync": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sync operations for given portfolio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc"
                ],
                "summary": "Sync operations",
                "operationId": "sync-op",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter operations from this date",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter operations till this date",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.commonResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.commonResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "api.getAverageSuccess": {
            "type": "object",
            "properties": {
                "avg": {
                    "type": "number",
                    "example": 42
                }
            }
        },
        "api.getBalanceSuccess": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 42
                }
            }
        },
        "api.gmailAuthURLSuccess": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://google.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}